+++ blocks/keyfeatures/keyfeatures.js
@@ -1,6 +1,6 @@
 import { ctaClick, ctaClickInteraction, keyFeaturesInteraction } from '../../dl.js';
 import { autoLinkLangPath, fetchPlaceholders } from '../../scripts/aem.js';
-import { decoratePlaceholder, renderHelper, targetObject } from '../../scripts/scripts.js';
+import { decoratePlaceholder, targetObject } from '../../scripts/scripts.js';
 import { renderHelper } from '../../scripts/common.js';
 
 export default async function decorate(block) {
